fix: fix remaining Windows compilation errors
- Cast handle to isize for HANDLE struct in windows_signals.rs
- Make info mutable in port.rs to allow field assignments
- Fix Windows-specific code that was only caught by Windows CI

+++ src/serial_core/windows_signals.rs
@@ -107,7 +107,7 @@ impl WindowsSignalControl {
         let file = self.open_port_handle()?;
         let handle = file.as_raw_handle();
 
-        unsafe { Self::set_dtr_handle(HANDLE(handle), enable) }
+        unsafe { Self::set_dtr_handle(HANDLE(handle as isize), enable) }
     }
 
     /// Set RTS signal by port name (convenience method)
@@ -119,7 +119,7 @@ impl WindowsSignalControl {
         let file = self.open_port_handle()?;
         let handle = file.as_raw_handle();
 
-        unsafe { Self::set_rts_handle(HANDLE(handle), enable) }
+        unsafe { Self::set_rts_handle(HANDLE(handle as isize), enable) }
     }
 }
